Alterando nome do controller de lista de itens para ItemController
